Ball of Mud Commit
Started adding ability to create a new user via the API. Was going to
commit once that was working. Realized too late that this is a
way-to-big commit. It's not working yet, but I want to record it at this
state, because I'm about to do some major refactoring, and I want a
place to back up to if it goes South.